tu ma

+++ src/com/girlkun/models/player/tuma/TuMaTemplate.java
@@ -17,11 +17,11 @@ public class TuMaTemplate {
     public static List<CongPhapTuMa> CONG_PHAP = new ArrayList<>();
 
     public void initLinhCan() {
-        LINH_CAN.add(new LinhCanTuMa("Huyết Linh Căn", "Linh căn khát máu hút #% HP của đối thủ chuyển thành của mình", (short) 0, (byte) 0));
-        LINH_CAN.add(new LinhCanTuMa("Sát Linh Căn", "Phóng thích sát khí gây sát thương chuẩn bằng #% máu của đối thủ", (short) 0, (byte) 1));
-        LINH_CAN.add(new LinhCanTuMa("Ma Linh Căn", "Đòn đánh thường gây thêm #% sát thương Ám ảnh", (short) 0, (byte) 2));
-        LINH_CAN.add(new LinhCanTuMa("Hắc Linh Căn", "Đòn đánh thường có #% gây mê muội và #% sát thương chuẩn", (short) 0, (byte) 3));
-        LINH_CAN.add(new LinhCanTuMa("Bạo Thực Linh Căn", "Linh căn cuồng bạo gây sát thương bằng #%Hp của đối thủ", (short) 0, (byte) 4));
+        LINH_CAN.add(new LinhCanTuMa("Huyết Linh Căn", "Linh căn khát máu hút #% HP của đối thủ chuyển thành của mình", 0, (byte) 0));
+        LINH_CAN.add(new LinhCanTuMa("Sát Linh Căn", "Phóng thích sát khí gây sát thương chuẩn bằng #% máu của đối thủ", 0, (byte) 1));
+        LINH_CAN.add(new LinhCanTuMa("Ma Linh Căn", "Đòn đánh thường gây thêm #% sát thương Ám ảnh", 0, (byte) 2));
+        LINH_CAN.add(new LinhCanTuMa("Hắc Linh Căn", "Đòn đánh thường có #% gây mê muội và #% sát thương chuẩn", 0, (byte) 3));
+        LINH_CAN.add(new LinhCanTuMa("Bạo Thực Linh Căn", "Linh căn cuồng bạo gây sát thương bằng #%Hp của đối thủ", 0, (byte) 4));
     }
 
     public void initCongPhap() {
